+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 Note: the BigBuy API is not versioned and some endpoints are often added or removed without notice.
 The removal of methods that call deleted endpoints is not considered a breaking change.
 
+## 3.23.0 (unreleased)
+
+* Remove `get_modules` (BigBuy removed the route)
+* Remove `get_module_platforms` (BigBuy now returns an empty response and the route was not documented anyway)
+* More functions use typed dicts in their return type: `get_order_statuses`, `get_product_compliance`,
+  `get_products_prices`, `get_products_tags`, `get_product_stock_by_handling_days`, `get_products_variations`,
+  `get_products_variations_stock_by_handling_days`, `get_product_tags`, `get_tag`, `get_tag_all_languages`, `get_tags`,
+  `get_products_taxonomies`, `get_carriers`
+* Add more type hints
+
 ## 3.22.0 (2025/04/14)
 
 ### Breaking Changes